project/requirements.txt
anyio==3.6.1
asgiref==3.5.2
black==22.6.0
click==8.1.3
fastapi==0.70.1
h11==0.13.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pydantic==1.9.1
sniffio==1.2.0
starlette==0.16.0
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.16.0
asyncpg==0.25.0
tortoise-orm==0.17.8
aerich==0.6.1
